requirements.txt
Django>=4.0.7
djangorestframework>=3.12.4
psycopg2==2.9.3
python-dotenv==0.20.0
drf-yasg==1.21.4
